+++ CHANGELOG.md
@@ -15,3 +15,8 @@ All notable changes to the "license-inserter" extension will be documented in th
 
 - Fix not honoring the extension settings
 - Format languages with /**/ block comment to prefix all lines in the header with *
+
+## [1.0.3]
+
+- Fix a bug that caused to fail if vscode language configuration is JSON with comments
+- Change invocation command from `Copyright: insert.header` to `Copyright: Insert Header`